E-board connection type change problem fixed.
Cause of the problem:
As soon as a device connection type is set to RS-232, an event handler
displays a combo box to display and/or accept the COM port ID for the device.
The handler would also tries to select one item of the combo box referring
parking lot database.
However, when COM_ID field is not initialized in the database,
the handler fails to make initial box item selection and the problem happens.

Solution:
In case COM_ID field is not initialized in the database, the event handler is
changed to choose the first item of the combo box which is COM1.