Array Index out of bounds exception fixed.
The exception happened under the following scenario.
A user starts creating a driver, then, he/she selects driver's
1st level affiliation, but don't select 2nd level affiliation.
At this point of time, if the user tries to save the new driver information,
OsParking responds with a confirm dialog. On the dialog if the user
decides not to select/provide 2nd level affiliation, the exception occurs.