Car entry button and Camera simulator's role clearly defined.
Here car entry button(ÀÔÂ÷¹öư: ·¹À̺íÀº ÀԱ¸¸í) is only for
gate #1 and the button is enabled only for 'admin' in the debug mode.
In the settings form, for gate #1, a new camera type option item
"CarEntry(=ÀÔÂ÷¹öư)" is added.
For gate 1, if car entry button is chosen to be used, camera
simulator #1 cannot be started.
A couple of shortcut key reassigned.